Reduce duplication and sanatize license
